- Fixed warning in ATerm library (a size_t was printed with %ld) - Fixed bug in communication calculation function of mcrl22lps; a   auxiliary function assumed sortedness of the arguments, but this was not   always the case.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2698 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/source/memory.c
@@ -460,7 +460,7 @@ void AT_initMemory(unsigned int argc, char *argv[])
               TERM_CHECK_OPT);
 #endif
       fprintf(stderr, "    %-20s: initial maximum term size " 
-              "(minimum=%ld, default=%d)\n", TERM_SIZE_OPT " <size>", MIN_TERM_SIZE, maxTermSize);
+              "(minimum=%d, default=%d)\n", TERM_SIZE_OPT " <size>", (unsigned int) MIN_TERM_SIZE, maxTermSize);
     } 
   }
